requirements.txt
Django==1.3.1
Flask==0.8
Flask-WTF==0.5.2
Jinja2==2.6
Markdown==2.1.1
PyRSS2Gen==1.0.0
PyYAML==3.10
Twisted==11.0.0
WTForms==0.6.3
Werkzeug==0.8.3
altgraph==0.7.1
amqplib==1.0.2
anyjson==0.3.1
bdist-mpkg==0.4.4
bonjour-py==0.3
celery==2.5.5
celery-with-redis==2.5
certifi==0.0.8
chardet==1.0.1
django-taggit==0.9.3
gevent==0.13.7
gevent-websocket==0.3.6
greenlet==0.4.0
include-server==3.1-toolwhip.1
kombu==2.1.8
lettuce==0.1.34
macholib==1.3
minimongo==0.2.7
modulegraph==0.8.1
nltk==2.0.1rc4
nose==1.1.2
numpy==1.5.1
py2app==0.5.3
pyOpenSSL==0.12
pymongo==2.1.1
pyobjc-core==2.3.2a0
pyobjc-framework-AddressBook==2.3.2a0
pyobjc-framework-AppleScriptKit==2.3.2a0
pyobjc-framework-AppleScriptObjC==2.3.2a0
pyobjc-framework-Automator==2.3.2a0
pyobjc-framework-CFNetwork==2.3.2a0
pyobjc-framework-CalendarStore==2.3.2a0
pyobjc-framework-Cocoa==2.3.2a0
pyobjc-framework-Collaboration==2.3.2a0
pyobjc-framework-CoreData==2.3.2a0
pyobjc-framework-CoreLocation==2.3.2a0
pyobjc-framework-CoreText==2.3.2a0
pyobjc-framework-DictionaryServices==2.3.2a0
pyobjc-framework-ExceptionHandling==2.3.2a0
pyobjc-framework-FSEvents==2.3.2a0
pyobjc-framework-InputMethodKit==2.3.2a0
pyobjc-framework-InstallerPlugins==2.3.2a0
pyobjc-framework-InstantMessage==2.3.2a0
pyobjc-framework-InterfaceBuilderKit==2.3.2a0
pyobjc-framework-LatentSemanticMapping==2.3.2a0
pyobjc-framework-LaunchServices==2.3.2a0
pyobjc-framework-Message==2.3.2a0
pyobjc-framework-OpenDirectory==2.3.2a0
pyobjc-framework-PreferencePanes==2.3.2a0
pyobjc-framework-PubSub==2.3.2a0
pyobjc-framework-QTKit==2.3.2a0
pyobjc-framework-Quartz==2.3.2a0
pyobjc-framework-ScreenSaver==2.3.2a0
pyobjc-framework-ScriptingBridge==2.3.2a0
pyobjc-framework-SearchKit==2.3.2a0
pyobjc-framework-ServerNotification==2.3.2a0
pyobjc-framework-ServiceManagement==2.3.2a0
pyobjc-framework-SyncServices==2.3.2a0
pyobjc-framework-SystemConfiguration==2.3.2a0
pyobjc-framework-WebKit==2.3.2a0
pyobjc-framework-XgridFoundation==2.3.2a0
python-dateutil==1.5
redis==2.4.13
requests==0.10.8
simplejson==2.4.0
slugify==0.0.1
virtualenv==1.7.1.2
virtualenv-clone==0.2.4
virtualenvwrapper==3.5
wsgiref==0.1.2
wxPython==2.8.12.1
wxPython-common==2.8.12.1
xattr==0.6.1
zope.interface==3.5.1
